rewrite static methods in shared class separately
